requirements.txt
PySide6>=6.6.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
cryptography>=42.0.0
pywin32>=306
